refactor(moq-mux): per-container subdirs, collapse import/ to import.rs
Reorganize so each external file format and each wire-level container has
its own module:

- container/fmp4/ — Cmaf (Container impl) + import.rs (Fmp4 → Import) +
  export.rs (Fmp4 → Export). Tests + test_data live here too.
- container/mkv/ — import.rs + export.rs + tests (no wire Container; MKV
  is external-only).
- container/loc/ — Container impl, wire-only.
- container/legacy/ — Container impl extracted from the old Hang::Legacy
  dispatch arm. Hang::Legacy now carries a `Legacy` value, matching the
  Cmaf and Loc variants.
- container/hls/ — import.rs (no exporter today).
- container/source.rs — ExportSource + CatalogSource (used by both
  fmp4 and mkv exporters).

The top-level export/ directory is gone; everything moved into container/.

src/import/ collapses to a single src/import.rs containing FramedFormat /
StreamFormat dispatchers. fMP4 test fixtures (bbb.mp4, av1.mp4, vp9.mp4)
move to container/fmp4/test_data/ next to the tests that use them.

External callers update to the new paths: moq_mux::container::fmp4::export::Export,
moq_mux::container::mkv::export::Export, etc.
