+++ rs/moq-mux/src/codec/aac/import.rs
@@ -40,7 +40,10 @@ impl Import {
 
 		Ok(Self {
 			catalog,
-			track: crate::container::Producer::new(track, crate::container::Hang::Legacy),
+			track: crate::container::Producer::new(
+				track,
+				crate::container::Hang::Legacy(crate::container::legacy::Legacy::new()),
+			),
 			zero: None,
 		})
 	}
