+++ rs/moq-mux/src/codec/av1/import.rs
@@ -111,7 +111,10 @@ impl Import {
 			.insert(track.name.clone(), config.clone());
 
 		self.config = Some(config);
-		self.track = Some(crate::container::Producer::new(track, crate::container::Hang::Legacy));
+		self.track = Some(crate::container::Producer::new(
+			track,
+			crate::container::Hang::Legacy(crate::container::legacy::Legacy::new()),
+		));
 
 		Ok(())
 	}
@@ -155,7 +158,10 @@ impl Import {
 			.insert(track.name.clone(), config.clone());
 
 		self.config = Some(config);
-		self.track = Some(crate::container::Producer::new(track, crate::container::Hang::Legacy));
+		self.track = Some(crate::container::Producer::new(
+			track,
+			crate::container::Hang::Legacy(crate::container::legacy::Legacy::new()),
+		));
 
 		Ok(())
 	}
@@ -244,7 +250,10 @@ impl Import {
 			.insert(track.name.clone(), config.clone());
 
 		self.config = Some(config);
-		self.track = Some(crate::container::Producer::new(track, crate::container::Hang::Legacy));
+		self.track = Some(crate::container::Producer::new(
+			track,
+			crate::container::Hang::Legacy(crate::container::legacy::Legacy::new()),
+		));
 
 		Ok(())
 	}
