+++ rs/moq-mux/src/codec/opus/import.rs
@@ -37,7 +37,10 @@ impl Import {
 
 		Ok(Self {
 			catalog,
-			track: crate::container::Producer::new(track, crate::container::Hang::Legacy),
+			track: crate::container::Producer::new(
+				track,
+				crate::container::Hang::Legacy(crate::container::legacy::Legacy::new()),
+			),
 			zero: None,
 		})
 	}
