+++ rs/moq-mux/src/container/consumer.rs
@@ -394,7 +394,7 @@ mod tests {
 		track.consume()
 	}
 
-	/// Write a finished group with explicit sequence and timestamps (Hang::Legacy format).
+	/// Write a finished group with explicit sequence and timestamps (Hang::Legacy(crate::container::legacy::Legacy::new()) format).
 	fn write_group(track: &mut moq_net::TrackProducer, sequence: u64, timestamps: &[Timestamp]) {
 		let mut group = track.create_group(moq_net::Group { sequence }).unwrap();
 		for &timestamp in timestamps {
@@ -403,7 +403,9 @@ mod tests {
 				payload: Bytes::from_static(&[0xDE, 0xAD]),
 				keyframe: false,
 			};
-			Hang::Legacy.write(&mut group, &[frame]).unwrap();
+			Hang::Legacy(crate::container::legacy::Legacy::new())
+				.write(&mut group, &[frame])
+				.unwrap();
 		}
 		group.finish().unwrap();
 	}
@@ -431,7 +433,8 @@ mod tests {
 	async fn read_single_group() {
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(500));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(500));
 
 		write_group(&mut track, 0, &[ts(0)]);
 		track.finish().unwrap();
@@ -449,7 +452,8 @@ mod tests {
 	async fn read_multiple_frames_single_group() {
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(500));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(500));
 
 		write_group(&mut track, 0, &[ts(0), ts(33_000), ts(66_000)]);
 		track.finish().unwrap();
@@ -467,7 +471,8 @@ mod tests {
 	async fn read_multiple_groups_within_latency() {
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(500));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(500));
 
 		// 5 groups, 20ms spacing. Total span = 80ms, well within 500ms latency.
 		for i in 0..5u64 {
@@ -486,12 +491,13 @@ mod tests {
 		tokio::time::pause();
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(100));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(100));
 
 		// Group 0: 5 frames, NOT finished (blocks consumer)
 		let mut group0 = track.create_group(moq_net::Group { sequence: 0 }).unwrap();
 		for f in 0..5u64 {
-			Hang::Legacy
+			Hang::Legacy(crate::container::legacy::Legacy::new())
 				.write(
 					&mut group0,
 					&[Frame {
@@ -527,10 +533,11 @@ mod tests {
 		tokio::time::pause();
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::ZERO);
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::ZERO);
 
 		let mut group0 = track.create_group(moq_net::Group { sequence: 0 }).unwrap();
-		Hang::Legacy
+		Hang::Legacy(crate::container::legacy::Legacy::new())
 			.write(
 				&mut group0,
 				&[Frame {
@@ -563,10 +570,11 @@ mod tests {
 		tokio::time::pause();
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(100));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(100));
 
 		let mut group0 = track.create_group(moq_net::Group { sequence: 0 }).unwrap();
-		Hang::Legacy
+		Hang::Legacy(crate::container::legacy::Legacy::new())
 			.write(
 				&mut group0,
 				&[Frame {
@@ -605,10 +613,11 @@ mod tests {
 		tokio::time::pause();
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(500));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(500));
 
 		let mut group0 = track.create_group(moq_net::Group { sequence: 0 }).unwrap();
-		Hang::Legacy
+		Hang::Legacy(crate::container::legacy::Legacy::new())
 			.write(
 				&mut group0,
 				&[Frame {
@@ -640,7 +649,8 @@ mod tests {
 	async fn adjacent_group_flushed_immediately() {
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(500));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(500));
 
 		write_group(&mut track, 0, &[ts(0)]);
 		write_group(&mut track, 1, &[ts(30_000)]);
@@ -658,7 +668,8 @@ mod tests {
 	async fn bframes_within_group() {
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(500));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(500));
 
 		write_group(&mut track, 0, &[ts(0), ts(66_000), ts(33_000)]);
 		track.finish().unwrap();
@@ -677,7 +688,8 @@ mod tests {
 		tokio::time::pause();
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(500));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(500));
 
 		track.finish().unwrap();
 
@@ -695,7 +707,8 @@ mod tests {
 		tokio::time::pause();
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(500));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(500));
 
 		write_group(&mut track, 0, &[ts(0)]);
 		track.abort(moq_net::Error::Cancel).unwrap();
@@ -717,7 +730,8 @@ mod tests {
 		tokio::time::pause();
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(500));
+		let consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(500));
 
 		assert!(
 			tokio::time::timeout(Duration::from_millis(50), consumer.closed())
@@ -740,7 +754,8 @@ mod tests {
 	async fn gap_in_group_sequence_recovery() {
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(100));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(100));
 
 		write_group(&mut track, 0, &[ts(0), ts(20_000)]);
 		write_group(&mut track, 1, &[ts(40_000), ts(60_000)]);
@@ -758,7 +773,8 @@ mod tests {
 	async fn gap_at_start_of_sequence() {
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(80));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(80));
 
 		write_group(&mut track, 5, &[ts(0), ts(20_000)]);
 		write_group(&mut track, 7, &[ts(80_000), ts(100_000)]);
@@ -776,7 +792,8 @@ mod tests {
 	async fn frame_timestamp_and_index_decoding() {
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(500));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(500));
 
 		write_group(&mut track, 0, &[ts(0), ts(33_333), ts(66_666)]);
 		track.finish().unwrap();
@@ -796,11 +813,12 @@ mod tests {
 	async fn frame_payload_preserved() {
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(500));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(500));
 
 		let payload_bytes = vec![0x01, 0x02, 0x03, 0x04, 0x05];
 		let mut group = track.create_group(moq_net::Group { sequence: 0 }).unwrap();
-		Hang::Legacy
+		Hang::Legacy(crate::container::legacy::Legacy::new())
 			.write(
 				&mut group,
 				&[Frame {
@@ -833,10 +851,11 @@ mod tests {
 		tokio::time::pause();
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_secs(10));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_secs(10));
 
 		let mut group0 = track.create_group(moq_net::Group { sequence: 0 }).unwrap();
-		Hang::Legacy
+		Hang::Legacy(crate::container::legacy::Legacy::new())
 			.write(
 				&mut group0,
 				&[Frame {
@@ -878,7 +897,8 @@ mod tests {
 	async fn large_timestamps() {
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_secs(3700));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_secs(3700));
 
 		let one_hour = 3_600_000_000u64;
 		write_group(&mut track, 0, &[ts(one_hour)]);
@@ -894,7 +914,8 @@ mod tests {
 	async fn set_latency_changes_behavior() {
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_secs(10));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_secs(10));
 
 		write_group(&mut track, 0, &[ts(0)]);
 		track.finish().unwrap();
@@ -914,11 +935,12 @@ mod tests {
 		let consumer_track = track.consume();
 		// latency must exceed (group1_max - group0_min) = 100ms - 0ms = 100ms
 		// to avoid the latency skip and test B-frame timestamp tracking.
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(110));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(110));
 
 		let mut group0 = track.create_group(moq_net::Group { sequence: 0 }).unwrap();
 		for &timestamp in &[ts(0), ts(66_000), ts(33_000)] {
-			Hang::Legacy
+			Hang::Legacy(crate::container::legacy::Legacy::new())
 				.write(
 					&mut group0,
 					&[Frame {
@@ -963,13 +985,14 @@ mod tests {
 		tokio::time::pause();
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(100));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(100));
 
 		write_group(&mut track, 3, &[ts(0)]);
 		write_group(&mut track, 5, &[ts(150_000)]);
 
 		let mut group7 = track.create_group(moq_net::Group { sequence: 7 }).unwrap();
-		Hang::Legacy
+		Hang::Legacy(crate::container::legacy::Legacy::new())
 			.write(
 				&mut group7,
 				&[Frame {
@@ -982,7 +1005,7 @@ mod tests {
 
 		let finisher = tokio::spawn(async move {
 			tokio::time::sleep(Duration::from_millis(50)).await;
-			Hang::Legacy
+			Hang::Legacy(crate::container::legacy::Legacy::new())
 				.write(
 					&mut group7,
 					&[Frame {
@@ -1014,7 +1037,8 @@ mod tests {
 		tokio::time::pause();
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(500));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(500));
 
 		let _group5 = track.create_group(moq_net::Group { sequence: 5 }).unwrap();
 		write_group(&mut track, 7, &[ts(210_000)]);
@@ -1037,7 +1061,8 @@ mod tests {
 	async fn startup_single_group_mid_stream() {
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(500));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(500));
 
 		write_group(&mut track, 100, &[ts(3_000_000)]);
 		track.finish().unwrap();
@@ -1051,10 +1076,11 @@ mod tests {
 		tokio::time::pause();
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(50));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(50));
 
 		let mut group0 = track.create_group(moq_net::Group { sequence: 0 }).unwrap();
-		Hang::Legacy
+		Hang::Legacy(crate::container::legacy::Legacy::new())
 			.write(
 				&mut group0,
 				&[Frame {
@@ -1086,10 +1112,11 @@ mod tests {
 		tokio::time::pause();
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(100));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(100));
 
 		let mut group0 = track.create_group(moq_net::Group { sequence: 0 }).unwrap();
-		Hang::Legacy
+		Hang::Legacy(crate::container::legacy::Legacy::new())
 			.write(
 				&mut group0,
 				&[Frame {
@@ -1123,11 +1150,12 @@ mod tests {
 		tokio::time::pause();
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(100));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(100));
 
 		// Group 0: stalled at ts=0, NOT finished
 		let mut group0 = track.create_group(moq_net::Group { sequence: 0 }).unwrap();
-		Hang::Legacy
+		Hang::Legacy(crate::container::legacy::Legacy::new())
 			.write(
 				&mut group0,
 				&[Frame {
@@ -1160,7 +1188,8 @@ mod tests {
 		tokio::time::pause();
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(100));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(100));
 
 		// Group 0: finished normally
 		write_group(&mut track, 0, &[ts(0), ts(20_000)]);
@@ -1176,7 +1205,8 @@ mod tests {
 	async fn group_error_skips_to_next() {
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(500));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(500));
 
 		let mut group0 = track.create_group(moq_net::Group { sequence: 0 }).unwrap();
 		group0.abort(moq_net::Error::Cancel).unwrap();
@@ -1193,7 +1223,8 @@ mod tests {
 		tokio::time::pause();
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(500));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(500));
 
 		write_group(&mut track, 0, &[ts(0)]);
 
@@ -1222,7 +1253,8 @@ mod tests {
 	async fn empty_group_advances() {
 		let mut track = moq_net::Track::new("test").produce();
 		let consumer_track = subscribe_default(&track);
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(500));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(500));
 
 		let mut group0 = track.create_group(moq_net::Group { sequence: 0 }).unwrap();
 		group0.finish().unwrap();
@@ -1242,9 +1274,10 @@ mod tests {
 
 		let mut track = moq_net::Track::new("video").produce();
 		let consumer_track = track.consume();
-		let mut consumer = Consumer::new(consumer_track, Hang::Legacy).with_latency(Duration::from_millis(500));
+		let mut consumer = Consumer::new(consumer_track, Hang::Legacy(crate::container::legacy::Legacy::new()))
+			.with_latency(Duration::from_millis(500));
 
-		// Write frames using Hang::Legacy encoding
+		// Write frames using Hang::Legacy(crate::container::legacy::Legacy::new()) encoding
 		let mut group = track.create_group(moq_net::Group { sequence: 0 }).unwrap();
 		for i in 0..3u64 {
 			let frame = Frame {
@@ -1252,7 +1285,9 @@ mod tests {
 				payload: Bytes::from_static(&[0xDE, 0xAD]),
 				keyframe: false,
 			};
-			Hang::Legacy.write(&mut group, &[frame]).unwrap();
+			Hang::Legacy(crate::container::legacy::Legacy::new())
+				.write(&mut group, &[frame])
+				.unwrap();
 		}
 		group.finish().unwrap();
 		track.finish().unwrap();
