+++ rs/moq-mux/src/container/fmp4/mod.rs
@@ -1,3 +1,19 @@
+//! fMP4 / CMAF: ISO-BMFF moof+mdat fragments.
+//!
+//! - [`Cmaf`] implements the wire-level [`crate::container::Container`] for a
+//!   single track (one moof+mdat per moq-lite frame, potentially N samples).
+//! - [`import::Import`] parses external fMP4 byte streams and publishes them.
+//! - [`export::Export`] subscribes to a broadcast and emits an fMP4 stream
+//!   (init segment + moof+mdat fragments).
+
+pub mod export;
+pub mod import;
+
+#[cfg(test)]
+mod export_test;
+#[cfg(test)]
+mod import_test;
+
 use std::task::Poll;
 
 use bytes::Bytes;
