+++ rs/moq-mux/src/container/hls/mod.rs
@@ -1,0 +1,8 @@
+//! HLS playlist ingest.
+//!
+//! HLS is an external streaming format only — no moq wire-level
+//! [`Container`] counterpart and no exporter today.
+//!
+//! [`Container`]: crate::container::Container
+
+pub mod import;
