--- rs/moq-mux/src/export/test/mod.rs
@@ -1,2 +1,0 @@
-mod fmp4;
-mod mkv;
